EGYCan's final website update
This branch (excluding the node_modules folder) contains all the new updates to the website. The website is built using EJS, a templating language used to make the development of websites modular and less sporadic. You can view a previous version of the website here on github via the link in the readme file. 

Current website will be live at egycan.com.eg